# server/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Codex Developer API",
    "description": "API endpoints that enable programmatic reading, writing, and updating of your Codex instance. UI supplied by Swagger.io."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/v1/workspace/{slug}/thread/new": {
      "post": {
        "tags": [
          "Workspace Threads"
        ],
        "description": "Create a new workspace thread",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique slug of workspace"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "thread": {
                      "id": 1,
                      "name": "Thread",
                      "slug": "thread-uuid",
                      "user_id": 1,
                      "workspace_id": 1
                    },
                    "message": null
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidAPIKey"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidAPIKey"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Optional userId associated with the thread, thread slug and thread name",
          "required": false,
          "content": {
            "application/json": {
              "example": {
                "userId": 1,
                "name": "Name",
                "slug": "thread-slug"
              }
            }
          }
        }
      }
    },
    "/v1/workspace/{slug}/thread/{threadSlug}/update": {
      "post": {
        "tags": [
          "Workspace Threads"
        ],
        "description": "Update thread name by its unique slug.",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique slug of workspace"
          },
          {
            "name": "threadSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique slug of thread"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "thread": {
                      "id": 1,
                      "name": "Updated Thread Name",
                      "slug": "thread-uuid",
                      "user_id": 1,
                      "workspace_id": 1
                    },
                    "message": null
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidAPIKey"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidAPIKey"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "JSON object containing new name to update the thread.",
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "name": "Updated Thread Name"
              }
            }
          }
        }
      }
    },
    "/v1/workspace/{slug}/thread/{threadSlug}": {
      "delete": {
        "tags": [
          "Workspace Threads"
        ],
        "description": "Delete a workspace thread",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique slug of workspace"
          },
          {
            "name": "threadSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique slug of thread"
          }
        ],
        "responses": {
          "200": {
            "description": "Thread deleted successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidAPIKey"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidAPIKey"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/v1/workspace/{slug}/thread/{threadSlug}/chats": {
      "get": {
        "tags": [
          "Workspace Threads"
        ],
        "description": "Get chats for a workspace thread",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique slug of workspace"
          },
          {
            "name": "threadSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique slug of thread"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "history": [
                      {
                        "role": "user",
                        "content": "What is Codex?",
                        "sentAt": 1692851630
                      },
                      {
                        "role": "assistant",
                        "content": "Codex is a platform that allows you to convert notes, PDFs, and other source materials into a chatbot. It ensures privacy, cites its answers, and allows multiple people to interact with the same documents simultaneously. It is particularly useful for businesses to enhance the visibility and readability of various written communications such as SOPs, contracts, and sales calls. You can try it out with a free trial to see if it meets your business needs.",
                        "sources": [
                          {
                            "source": "object about source document and snippets used"
                          }
                        ]
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidAPIKey"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidAPIKey"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/v1/workspace/{slug}/thread/{threadSlug}/chat": {
      "post": {
        "tags": [
          "Workspace Threads"
        ],
        "description": "Chat with a workspace thread",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique slug of workspace"
          },
          {
            "name": "threadSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique slug of thread"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "id": "chat-uuid",
                    "type": "abort | textResponse",
                    "textResponse": "Response to your query",
                    "sources": [
                      {
                        "title": "codex.txt",
                        "chunk": "This is a context chunk used in the answer of the prompt by the LLM."
                      }
                    ],
                    "close": true,
                    "error": "null | text string of the failure mode."
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidAPIKey"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidAPIKey"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Send a prompt to the workspace thread and the type of conversation (query or chat).",
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "message": "What is Codex?",
                "mode": "query | chat",
                "userId": 1,
                "attachments": [
                  {
                    "name": "image.png",
                    "mime": "image/png",
                    "contentString": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
                  }
                ],
                "reset": false
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "InvalidAPIKey": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Invalid API Key"
          }
        },
        "xml": {
          "name": "InvalidAPIKey"
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}